fixed list

+++ model/src/main/java/com/aday/model/room/BookListEntity.kt
@@ -9,6 +9,6 @@ data class BookListEntity(@PrimaryKey(autoGenerate = true) var id: Int?,
                           @ColumnInfo(name = "date") var date: String,
                           @ColumnInfo(name = "book_image") var image: String,
                           @ColumnInfo(name = "book_name") var name: String,
-                          @ColumnInfo(name = "rating") var rating: Int,
+                          @ColumnInfo(name = "rating") var rating: Float,
                           @ColumnInfo(name = "five_words") var fiveWords: String?,
                           @ColumnInfo(name = "detail_memo") var memo: String?)
